+++ src/components/Form/upload/atoms/textarea/UploadTextarea.tsx
@@ -15,7 +15,7 @@ function UploadTextarea({
   value,
 }: InputProps) {
   return (
-    <InputWrapper>
+    <InputContainer>
       <Label htmlFor={labelText}>
         <span>{labelText}</span>
         {maxLength && <span>{`${value.length} / ${maxLength}`}</span>}
@@ -27,7 +27,7 @@ function UploadTextarea({
         onChange={(event) => onChangeInputValue(event.target.value)}
         rows={6}
       />
-    </InputWrapper>
+    </InputContainer>
   );
 }
 
@@ -37,7 +37,7 @@ const Label = styled.label`
   margin-bottom: 0.3rem;
 `;
 
-const InputWrapper = styled.section`
+const InputContainer = styled.section`
   display: flex;
   flex-direction: column;
 `;
